refactor: Canvas painter component mechanism
